Move StacClient Logging mid into the stac module
