+++ stac-example/src/main/resources/logback.xml
@@ -12,8 +12,8 @@
   <logger name="org.http4s.blaze.channel.nio1" level="${BLAZE_NIO_LOG_LEVEL:-WARN}"/>
   <logger name="software.amazon.awssdk" level="${AWS_SDK_LOG_LEVEL:-WARN}"/>
   <logger name="geotrellis.raster.effects" level="${GT_MOSAIC_LOG_LEVEL:-DEBUG}" />
-  <logger name="geotrellis.server.ogc.stac.util.logging.StacClientLoggingMid" level="${GT_STAC_CLIENT_LOG_LEVEL:-INFO}"/>
-  <logger name="geotrellis.server.ogc.stac.util.logging.StreamingStacClientLoggingMid" level="${GT_STREAMING_STAC_CLIENT_LOG_LEVEL:-INFO}"/>
+  <logger name="geotrellis.stac.util.logging.StacClientLoggingMid" level="${GT_STAC_CLIENT_LOG_LEVEL:-INFO}"/>
+  <logger name="geotrellis.stac.util.logging.StreamingStacClientLoggingMid" level="${GT_STREAMING_STAC_CLIENT_LOG_LEVEL:-INFO}"/>
 
   <root level="DEBUG">
     <appender-ref ref="STDOUT" />
